+++ DataLayer/Event/Purchase.php
@@ -8,6 +8,7 @@ use Magento\Sales\Model\Order;
 use Yireo\GoogleTagManager2\Api\Data\EventInterface;
 use Yireo\GoogleTagManager2\Config\Config;
 use Yireo\GoogleTagManager2\DataLayer\Tag\Order\OrderItems;
+use Yireo\GoogleTagManager2\Util\OrderTotals;
 use Yireo\GoogleTagManager2\Util\PriceFormatter;
 
 // See https://developers.google.com/analytics/devguides/collection/ga4/reference/events?client_type=gtm#purchase
@@ -18,17 +19,20 @@ class Purchase implements EventInterface
     private Config $config;
     private PriceFormatter $priceFormatter;
     private Session $checkoutSession;
+    private OrderTotals $orderTotals;
 
     public function __construct(
         OrderItems $orderItems,
         Config $config,
         PriceFormatter $priceFormatter,
-        Session $checkoutSession
+        Session $checkoutSession,
+        OrderTotals $orderTotals
     ) {
         $this->orderItems = $orderItems;
         $this->config = $config;
         $this->priceFormatter = $priceFormatter;
         $this->checkoutSession = $checkoutSession;
+        $this->orderTotals = $orderTotals;
     }
 
     /**
@@ -53,7 +57,7 @@ class Purchase implements EventInterface
                 'currency' => $order->getOrderCurrencyCode(),
                 'value' => $this->priceFormatter->format((float)$order->getSubtotal()),
                 'tax' => $this->priceFormatter->format((float)$order->getTaxAmount()),
-                'shipping' => $this->priceFormatter->format((float)$order->getShippingAmount()),
+                'shipping' => $this->priceFormatter->format($this->orderTotals->getShippingTotal($order)),
                 'coupon' => $order->getCouponCode(),
                 'payment_method' => $order->getPayment() ? $order->getPayment()->getMethod() : '',
                 'items' => $this->orderItems->setOrder($order)->get(),