+++ DataLayer/Event/Refund.php
@@ -6,6 +6,7 @@ use Magento\Sales\Api\Data\OrderInterface;
 use Yireo\GoogleTagManager2\Api\Data\EventInterface;
 use Yireo\GoogleTagManager2\Config\Config;
 use Yireo\GoogleTagManager2\DataLayer\Tag\Order\OrderItems;
+use Yireo\GoogleTagManager2\Util\OrderTotals;
 use Yireo\GoogleTagManager2\Util\PriceFormatter;
 
 // See https://developers.google.com/analytics/devguides/collection/ga4/reference/events?client_type=gtm#refund
@@ -15,15 +16,18 @@ class Refund implements EventInterface
     private OrderItems $orderItems;
     private Config $config;
     private PriceFormatter $priceFormatter;
+    private OrderTotals $orderTotals;
 
     public function __construct(
         OrderItems $orderItems,
         Config $config,
-        PriceFormatter $priceFormatter
+        PriceFormatter $priceFormatter,
+        OrderTotals $orderTotals
     ) {
         $this->orderItems = $orderItems;
         $this->config = $config;
         $this->priceFormatter = $priceFormatter;
+        $this->orderTotals = $orderTotals;
     }
 
     /**
@@ -40,7 +44,7 @@ class Refund implements EventInterface
                 'currency' => $order->getOrderCurrencyCode(),
                 'value' => $this->priceFormatter->format((float)$order->getSubtotal()),
                 'tax' => $this->priceFormatter->format((float)$order->getTaxAmount()),
-                'shipping' => $this->priceFormatter->format((float)$order->getShippingAmount()),
+                'shipping' => $this->priceFormatter->format($this->orderTotals->getShippingTotal($order)),
                 'coupon' => $order->getCouponCode(),
                 'payment_method' => $order->getPayment() ? $order->getPayment()->getMethod() : '',
                 'items' => $this->orderItems->setOrder($order)->get(),